fixed code style; fixed IllegalArgumentException in AccountService; added test with empty id for AccountService

+++ aspsp-mock-server/src/main/java/de/adorsys/aspsp/aspspmockserver/service/AccountService.java
@@ -29,7 +29,8 @@ public class AccountService {
     }
 
     public boolean deleteAccountById(String id) {
-        if(accountRepository.exists(id)) {
+        if(id!=null&&accountRepository.exists(id))
+        {
             accountRepository.delete(id);
             return true;
         }
